# Some clean-up of remez