Start of an attempt to make this more readable. I found it quite hard to keep track of what all those variables mean, so I've started eliminating some of them.

Also, there is one algorithmic change, basically doing `grid .= cospi.(2.*grid)`. Before, the cosine might have potentially been applied only to a subset of grid points, but OTOH potentially multiple times. Net performance impact seems small (but an improvement), but this allows to remove some `cospi` calls, making for slightly less code.

What does have a significant performance impact is fixing a type-instabilty in `build_grid`:
```julia
# master
julia> @btime remez(20, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
  283.074 μs (8114 allocations: 142.48 KiB)

julia> @btime remez(151, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
  4.846 ms (87515 allocations: 1.41 MiB)
```
```julia
# this PR
julia> @btime remez(20, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
  53.988 μs (88 allocations: 11.80 KiB)

julia> @btime remez(151, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
  2.803 ms (104 allocations: 60.39 KiB)
```
@sirtom67 have you done the benchmarks shown at JuliaCon with the version from master? That might explain (part of) the overhead for small filter sizes.