@sirtom67 2018-08-24
Martin: Nice work! I will regenerate the timing plot with your fix. A
couple Julia experts commented at the conference that the slow speed for
small filter sizes was probably due to type instability. I had gotten
code_typewarn to run and and it found a number of things, but i have not
fixed or measured anything.

On Fri, Aug 24, 2018, 9:18 AM Martin Holters <notifications@github.com>
wrote:

> Start of an attempt to make this more readable. I found it quite hard to
> keep track of what all those variables mean, so I've started eliminating
> some of them.
>
> Also, there is one algorithmic change, basically doing grid .=
> cospi.(2.*grid). Before, the cosine might have potentially been applied
> only to a subset of grid points, but OTOH potentially multiple times. Net
> performance impact seems small (but an improvement), but this allows to
> remove some cospi calls, making for slightly less code.
>
> What does have a significant performance impact is fixing a
> type-instabilty in build_grid:
>
> # master
> julia> @btime remez(20, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
>   283.074 μs (8114 allocations: 142.48 KiB)
>
> julia> @btime remez(151, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
>   4.846 ms (87515 allocations: 1.41 MiB)
>
> # this PR
> julia> @btime remez(20, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
>   53.988 μs (88 allocations: 11.80 KiB)
>
> julia> @btime remez(151, $[0, 0.475, 0.5, 1.0], $[1.0, 0.0]; Hz=2.0);
>   2.803 ms (104 allocations: 60.39 KiB)
>
> @sirtom67 <https://github.com/sirtom67> have you done the benchmarks
> shown at JuliaCon with the version from master? That might explain (part
> of) the overhead for small filter sizes.
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/JuliaDSP/DSP.jl/pull/223
> Commit Summary
>
>    - Clean up `freq_eval` interface
>    - remez: Don't unnecessarily special-case last iteration of a loop
>    - remez: choose array lengths based on `nfcns`, eliminate `dimsize`
>    - remez: Pre-compute cospi(2grid)
>    - Move transformation of `des` and `at` into `build_grid`
>    - remez: some clean-up of variables
>    - Fix type-stability issue in build_grid
>
> File Changes
>
>    - *M* src/Filters/remez_fir.jl
>    <https://github.com/JuliaDSP/DSP.jl/pull/223/files#diff-0> (156)
>
> Patch Links:
>
>    - https://github.com/JuliaDSP/DSP.jl/pull/223.patch
>    - https://github.com/JuliaDSP/DSP.jl/pull/223.diff
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/JuliaDSP/DSP.jl/pull/223>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ALBY_4TcgwmZgdspJMjiVQh3Mt8rxcRnks5uUAsogaJpZM4WLb1z>
> .
>