@sirtom67 2018-09-01
I know this is merged already but I finally got around to pulling the latest on master and re-running my timing comparison script. Here's the plot:
![remez_time_julia_c_compare](https://user-images.githubusercontent.com/11557119/44950759-015a2a80-ae16-11e8-8fda-dc67fd01bea8.png)
As you can see the new version is much faster for small filters, and even faster than C by a bit for most orders.