@sirtom67 2018-09-01
Here's the old timing, for reference.
![remez_time_julia_c_compare-180729](https://user-images.githubusercontent.com/11557119/44950802-0f5c7b00-ae17-11e8-83b3-45c486d4831c.png)